カートページ作成

+++ database/migrations/2020_07_29_075722_create_carts_table.php
@@ -15,8 +15,8 @@ class CreateCartsTable extends Migration
     {
         Schema::create('carts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('stock_id');
-            $table->integer('user_id');
+            $table->bigInteger('stock_id');
+            $table->bigInteger('user_id');
             $table->timestamps();
         });
     }
